// frontend/src/apisv2/run/models/V2RuntimeState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * backend/api/v2/run.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Describes the runtime state of an entity.
 *
 *  - RUNTIME_STATE_UNSPECIFIED: Default value. This value is not used.
 *  - PENDING: Service is preparing to execute an entity.
 *  - RUNNING: Entity execution is in progress.
 *  - SUCCEEDED: Entity completed successfully.
 *  - SKIPPED: Entity has been skipped. For example, due to caching.
 *  - FAILED: Entity execution has failed.
 *  - CANCELING: Entity is being canceled. From this state, an entity may only
 * change its state to SUCCEEDED, FAILED or CANCELED.
 *  - CANCELED: Entity has been canceled.
 *  - PAUSED: Entity has been paused. It can be resumed.
 * @export
 */
export const V2RuntimeState = {
  RUNTIME_STATE_UNSPECIFIED: 'RUNTIME_STATE_UNSPECIFIED',
  PENDING: 'PENDING',
  RUNNING: 'RUNNING',
  SUCCEEDED: 'SUCCEEDED',
  SKIPPED: 'SKIPPED',
  FAILED: 'FAILED',
  CANCELING: 'CANCELING',
  CANCELED: 'CANCELED',
  PAUSED: 'PAUSED',
} as const;
export type V2RuntimeState = (typeof V2RuntimeState)[keyof typeof V2RuntimeState];

export function instanceOfV2RuntimeState(value: any): boolean {
  for (const key in V2RuntimeState) {
    if (Object.prototype.hasOwnProperty.call(V2RuntimeState, key)) {
      if (V2RuntimeState[key as keyof typeof V2RuntimeState] === value) {
        return true;
      }
    }
  }
  return false;
}

export function V2RuntimeStateFromJSON(json: any): V2RuntimeState {
  return V2RuntimeStateFromJSONTyped(json, false);
}

export function V2RuntimeStateFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): V2RuntimeState {
  return json as V2RuntimeState;
}

export function V2RuntimeStateToJSON(value?: V2RuntimeState | null): any {
  return value as any;
}

export function V2RuntimeStateToJSONTyped(
  value: any,
  ignoreDiscriminator: boolean,
): V2RuntimeState {
  return value as V2RuntimeState;
}